libusrxml: Overwrite file contents in store_usrxml_file()
It is assumed that old data in existing file will be substituded with
new one on store, not appended as that might make content inconsistent.

For example user entry might be deleted and thus missing in new data
while old data we append to still has this entry making it to reappear
on next data (re)load.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>