libusrxml: Link <user> and <interface> with lower/upper dependency
User is similar to network interface without any uppers and exactly one
lower interface that it is configured for. This is pure lower/upper
dependency and can be handled by existing network interface specific
code.

Since then, <if> in <user> has same meaning as <lower> in <interface>
and <user> is used instead of <upper> in <interface> when it points
to user instead of network interface entry.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>