libusrxml: Reorder arguments to callback function in run_usrxml_parser()
With commit 7c0f28d15e90 ("libusrxml: Return saved user id with handle
on delete") arguments order preserved to keep compatibility with
existing implementations.

Since that time there is a lot of changes made to codebase and there is
no prove that compatibility is kept. Thus change argument order to group
ret and handle with userid passed in array together.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>