libusrxml: Introduce helpers to iterate over dynamic map in reverse order
Base implementation on usrxml___dyn_for_each() to avoid code duplication
and make it true private, implementing code functionality, function.

To make code distinguish whenever we should traverse forward or backward
introduce new parameter @dec that is set to zero in forward case and to
non-zero in backward case.

Pass this new parameter as -1 in forward and 1 in backward cases to
callback function to let it known order in which data is traversed.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>