libusrxml: Delete destination dynamic map entries before copying source
Now we copy dynamic map entries for both user and interface, but without
removing destination entries if they existed. This could leave existing
entries in destination that isn't present in source.

To avoid this previously we used hack when restoring data that removes
new entries that not overwritten on copy. However this hack only covers
subset of possible cases and generally entry copy routines should care
about all these cases.

While there fix pipe "opts" deletion and pass @sh instead of @h to
usrxml__type_is_user() for clarity, even if it is not used in presence
of non-empty type.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>