libusrxml: Support dynamic map iterations starting from specific index
It could be useful to resume traversal of dynamic map starting from
specific index either in same or different direction.

For example apply specific @cb in forward direction until it failed
and then go in reverse with another @cb (or same that takes @dec in
account for direction) to rollback modifications.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>