libusrxml: Pass original @data from run_usrxml_parser() to ifupdown func
We might need to pass additional data to @fn called from ifupdown
callback processing USRXML_ifupdown list.

There are two formats of data array:

  1) Hierarcical where @fn stored under "usrxml","ifupdown"
  2) Plain where @fn stored directly in @data array.

Both formats supported with preference to first.

Note that it is not strictly necessary to known in @fn if we perform
normal up/down or rollback: thus we not copying "rb" from a[] to data[].

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>