libusrxml: Fix ifupdown callback in case of rollback
That callback determines if running rollback by checking "rb" presence
in array @a which is not passed anymore since we pass original @data
from run_usrxml_parser() to ifupdown callback.

Make @a part of @data array supplied to run_usrxml_parser() to pass
both user supplied @data and @a to the ifupdown callback and user @fn.

Also respect that @cb and/or @data of run_usrxml_parser() are optional
thus use local, empty @b array when @data not specified, prepare "a"
subarray in it and do that only when @cb is given.

Fixes: 1d032a8db460 ("libusrxml: Pass original @data from run_usrxml_parser() to ifupdown func")
Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>