libusrxml: Make user (de)activation helpers reentrant
The only thing is missing is userid filtering from compatibility dval in
dynamic USRXML_ifuser[] to avoid userid duplication on multiple calls
for same user.

While there get rid of unused local variable @ret and replace it with
new local variable @t that is also used as temporary storage for regexp
for gensub().

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>